LayoutTests/inspector/debugger/stepping/stepOver-expected.txt

Checking pause locations when stepping with "stepOver".


== Running test suite: Debugger.stepOver
-- Running test case: Debugger.stepOver.statements
PAUSED (debugger-statement)
PAUSE AT testStatements:12:5
      8    }
      9    
     10    function testStatements() {
 ->  11        |debugger;
     12        let x = 1;
     13        let y = 2;
     14    }

PAUSE AT testStatements:13:5
      9    
     10    function testStatements() {
     11        debugger;
 ->  12        |let x = 1;
     13        let y = 2;
     14    }
     15    

PAUSE AT testStatements:14:5
     10    function testStatements() {
     11        debugger;
     12        let x = 1;
 ->  13        |let y = 2;
     14    }
     15    
     16    function testFunctions() {

PAUSE AT testStatements:15:2
     11        debugger;
     12        let x = 1;
     13        let y = 2;
 ->  14    }|
     15    
     16    function testFunctions() {
     17        debugger;

RESUMED

-- Running test case: Debugger.stepOver.function
PAUSED (debugger-statement)
PAUSE AT testFunctions:18:5
     14    }
     15    
     16    function testFunctions() {
 ->  17        |debugger;
     18        let before = 1;
     19        a();
     20        let after = 2;

PAUSE AT testFunctions:19:5
     15    
     16    function testFunctions() {
     17        debugger;
 ->  18        |let before = 1;
     19        a();
     20        let after = 2;
     21    }

PAUSE AT testFunctions:20:5
     16    function testFunctions() {
     17        debugger;
     18        let before = 1;
 ->  19        |a();
     20        let after = 2;
     21    }
     22    

PAUSE AT testFunctions:21:5
     17        debugger;
     18        let before = 1;
     19        a();
 ->  20        |let after = 2;
     21    }
     22    
     23    function testEval() {

PAUSE AT testFunctions:22:2
     18        let before = 1;
     19        a();
     20        let after = 2;
 ->  21    }|
     22    
     23    function testEval() {
     24        debugger;

RESUMED

-- Running test case: Debugger.stepOver.eval
PAUSED (debugger-statement)
PAUSE AT testEval:25:5
     21    }
     22    
     23    function testEval() {
 ->  24        |debugger;
     25        let before = 1;
     26        eval("1 + 1");
     27        let after = 2;

PAUSE AT testEval:26:5
     22    
     23    function testEval() {
     24        debugger;
 ->  25        |let before = 1;
     26        eval("1 + 1");
     27        let after = 2;
     28    }

PAUSE AT testEval:27:5
     23    function testEval() {
     24        debugger;
     25        let before = 1;
 ->  26        |eval("1 + 1");
     27        let after = 2;
     28    }
     29    

PAUSE AT testEval:28:5
     24        debugger;
     25        let before = 1;
     26        eval("1 + 1");
 ->  27        |let after = 2;
     28    }
     29    
     30    function testLocalFunction() {

PAUSE AT testEval:29:2
     25        let before = 1;
     26        eval("1 + 1");
     27        let after = 2;
 ->  28    }|
     29    
     30    function testLocalFunction() {
     31        (function() {

RESUMED

-- Running test case: Debugger.stepOver.innerFunction
PAUSED (debugger-statement)
PAUSE AT <anonymous>:33:9
     29    
     30    function testLocalFunction() {
     31        (function() {
 ->  32            |debugger;
     33            let inner = 1;
     34        })();
     35        let outer = 2;

PAUSE AT <anonymous>:34:9
     30    function testLocalFunction() {
     31        (function() {
     32            debugger;
 ->  33            |let inner = 1;
     34        })();
     35        let outer = 2;
     36    }

PAUSE AT <anonymous>:35:6
     31        (function() {
     32            debugger;
     33            let inner = 1;
 ->  34        }|)();
     35        let outer = 2;
     36    }
     37    

PAUSE AT testLocalFunction:36:5
     32            debugger;
     33            let inner = 1;
     34        })();
 ->  35        |let outer = 2;
     36    }
     37    
     38    function testCommas() {

PAUSE AT testLocalFunction:37:2
     33            let inner = 1;
     34        })();
     35        let outer = 2;
 ->  36    }|
     37    
     38    function testCommas() {
     39        debugger;

RESUMED

-- Running test case: Debugger.stepOver.commas
PAUSED (debugger-statement)
PAUSE AT testCommas:40:5
     36    }
     37    
     38    function testCommas() {
 ->  39        |debugger;
     40        let x = 1,
     41            y = 2,
     42            z = 3;

PAUSE AT testCommas:41:5
     37    
     38    function testCommas() {
     39        debugger;
 ->  40        |let x = 1,
     41            y = 2,
     42            z = 3;
     43        a(), b(), c();

PAUSE AT testCommas:42:9
     38    function testCommas() {
     39        debugger;
     40        let x = 1,
 ->  41            |y = 2,
     42            z = 3;
     43        a(), b(), c();
     44        true && (a(), b(), c());

PAUSE AT testCommas:43:9
     39        debugger;
     40        let x = 1,
     41            y = 2,
 ->  42            |z = 3;
     43        a(), b(), c();
     44        true && (a(), b(), c());
     45    }

PAUSE AT testCommas:44:5
     40        let x = 1,
     41            y = 2,
     42            z = 3;
 ->  43        |a(), b(), c();
     44        true && (a(), b(), c());
     45    }
     46    

PAUSE AT testCommas:44:10
     40        let x = 1,
     41            y = 2,
     42            z = 3;
 ->  43        a(), |b(), c();
     44        true && (a(), b(), c());
     45    }
     46    

PAUSE AT testCommas:44:15
     40        let x = 1,
     41            y = 2,
     42            z = 3;
 ->  43        a(), b(), |c();
     44        true && (a(), b(), c());
     45    }
     46    

PAUSE AT testCommas:45:5
     41            y = 2,
     42            z = 3;
     43        a(), b(), c();
 ->  44        |true && (a(), b(), c());
     45    }
     46    
     47    function a() { }

PAUSE AT testCommas:45:14
     41            y = 2,
     42            z = 3;
     43        a(), b(), c();
 ->  44        true && (|a(), b(), c());
     45    }
     46    
     47    function a() { }

PAUSE AT testCommas:45:19
     41            y = 2,
     42            z = 3;
     43        a(), b(), c();
 ->  44        true && (a(), |b(), c());
     45    }
     46    
     47    function a() { }

PAUSE AT testCommas:45:24
     41            y = 2,
     42            z = 3;
     43        a(), b(), c();
 ->  44        true && (a(), b(), |c());
     45    }
     46    
     47    function a() { }

PAUSE AT testCommas:46:2
     42            z = 3;
     43        a(), b(), c();
     44        true && (a(), b(), c());
 ->  45    }|
     46    
     47    function a() { }
     48    function b() { }

RESUMED
